refactored cnn and rnn model + data collection for ABCDEMN
