+++ pyproject.toml
@@ -8,11 +8,15 @@ readme = "README.md"
 [tool.poetry.dependencies]
 python = "^3.10"
 opencv-python = "^4.10.0.84"
-tensorflow = "^2.17.0"
 mediapipe = "^0.10.15"
 pandas = "^2.2.3"
 scikit-learn = "^1.5.2"
 seaborn = "^0.13.2"
+pynput = "^1.7.7"
+nvidia-cudnn-cu11 = "8.6.0.163"
+nvidia-cuda-runtime-cu11 = "11.8.89"
+nvidia-cuda-nvcc-cu11 = "11.8.89"
+tensorflow = "^2.17.0"
 
 
 [tool.poetry.group.dev.dependencies]
